Coerces statquote objects to strings suitable for LaTeX

This function coerces statquote objects to strings suitable for rendering in LaTeX. Quotes and (potential LaTeX) sources are placed within suitable "epigraph" output format via the sprintf function.

Usage

as.latex(quotes, form = "\\epigraph{%s}{%s}\n\n", cite = TRUE)

Arguments

quotes

an object of class statquote returned from functions such as search_quotes or statquote

form

structure of the LaTeX output for the text (first argument) and source (second argument) passed to sprintf

cite

logical; should the cite field be included in the source output?

Value

character vector of formatted LaTeX quotes

See also

Author

Phil Chalmers

Examples

ll <- search_quotes("Tukey")
as.latex(ll)
#>  [1] "\\epigraph{If data analysis is to be well done, much of it must be a matter of judgment, and 'theory' whether statistical or non-statistical, will have to guide, not command.}{John Tukey, The Future of Data Analysis, Annals of Mathematical Statistics, 1962, Vol. 33 (1), 1962.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#>  [2] "\\epigraph{The physical sciences are used to 'praying over' their data, examining the same data from a variety of points of view. This process has been very rewarding, and has led to many extremely valuable insights. Without this sort of flexibility, progress in physical science would have been much slower. Flexibility in analysis is often to be had honestly at the price of a willingness not to demand that what has already been observed shall establish, or prove, what analysis suggests. In physical science generally, the results of praying over the data are thought of as something to be put to further test in another experiment, as indications rather than conclusions.}{John Tukey, The Future of Data Analysis, Annals of Mathematical Statistics, 1962, Vol. 33 (1), 1962.}\n\n"
#>  [3] "\\epigraph{If one technique of data analysis were to be exalted above all others for its ability to be revealing to the mind in connection with each of many different models, there is little doubt which one would be chosen. The simple graph has brought more information to the data analyst's mind than any other device. It specializes in providing indications of unexpected phenomena.}{John Tukey, The Future of Data Analysis,  The Annals of Mathematical Statistics, 1962, Vol. 33, No. 1. Page 49.}\n\n"                                                                                                                                                                                                                                                                                         
#>  [4] "\\epigraph{The most important maxim for data analysis to heed...is this: Far better an approximate answer to the right question, which is often vague, than an exact answer to the wrong question, which can always be made precise.}{John Tukey, The Future of Data Analysis,  The Annals of Mathematical Statistics, 1962, Vol. 33, No. 1. Page 13-14.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [5] "\\epigraph{The greatest possibilities of visual display lie in vividness and inescapability of the intended message. A visual display can stop your mental flow in its tracks and make you think. A visual display can force you to notice what you never expected to see.}{John Tukey, NA}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#>  [6] "\\epigraph{The purpose of [data] display is comparison (recognition of phenomena), not numbers ... The phenomena are the main actors, numbers are the supporting cast.}{John Tukey, NA}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>  [7] "\\epigraph{...But it is not always clear \\emph{which} 1000 words.}{John Tukey, 1973, NA}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#>  [8] "\\epigraph{The best thing about being a statistician is that you get to play in everyone's backyard.}{John Tukey, NA}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#>  [9] "\\epigraph{The worst, i.e., most dangerous, feature of 'accepting the null hypothesis' is the giving up of explicit uncertainty ... Mathematics can sometimes be put in such black-and-white terms, but our knowledge or belief about the external world never can.\ndoi:10.1214/ss/1177011945}{John Tukey, The Philosophy of Multiple Comparisons, Statistical Science, 1991, 6, 100-116.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                
#> [10] "\\epigraph{Statisticians classically asked the wrong question--and were willing to answer with a lie, one that was often a downright lie. They asked``Are the effects of A and B different?\" and they were willing to answer``no\". All we know about the world teaches us that the effects of A and B are always different--in some decimal place--for every A and B. Thus asking``Are the effects different?\" is foolish. What we should be answering first is``Can we tell the direction in which the effects of A differ from the effects of B?\" In other words, can we be confident about the direction from A to B? Is it``up\",``down\" or``uncertain\"?\ndoi:10.1214/ss/1177011945}{John Tukey, The Philosophy of Multiple Comparisons, Statistical Science, 1991, 6, 100-116.}\n\n"                 
#> [11] "\\epigraph{Empirical knowledge is always fuzzy! And theoretical knowledge, like all the laws of physics, as of today's date, is always wrong-in detail, though possibly providing some very good approximations indeed.\ndoi:10.1214/ss/1177011945}{John Tukey, The Philosophy of Multiple Comparisons, Statistical Science, 1991, 6, 100-116.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [12] "\\epigraph{The combination of some data and an aching desire for an answer does not ensure that a reasonable answer can be extracted from a given body of data.}{John W. Tukey, Sunset Salvo, The American Statistician Vol. 40 (1), 1986.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [13] "\\epigraph{The practical power of a statistical test is the product of its' statistical power and the probability of use.}{John W. Tukey, A Quick, Compact, Two Sample Test to Duckworth's Specifications}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [14] "\\epigraph{Since the aim of exploratory data analysis is to learn what seems to be, it should be no surprise that pictures play a vital role in doing it well.}{John W. Tukey, John W. Tukey's Works on Interactive Graphics.  The Annals of Statistics Vol. 30, No. 6 (Dec., 2002), pp. 1629-1639}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> [15] "\\epigraph{There is nothing better than a picture for making you think of questions you had forgotten to ask (even mentally).}{John W. Tukey & Paul Tukey, John W. Tukey's Works on Interactive Graphics.  The Annals of Statistics Vol. 30, No. 6 (Dec., 2002), pp. 1629-1639}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [16] "\\epigraph{The greatest value of a picture is when it forces us to notice what we never expected to see.}{John Tukey, Exploratory Data Analysis, 1977, p. vi}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [17] "\\epigraph{Exploratory data analysis can never be the whole story, but nothing else can serve as the foundation stone -- as the first step.}{John Tukey, Exploratory Data Analysis, 1977, p.3.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [18] "\\epigraph{Unless exploratory data analysis uncovers indications, usually quantitative ones, there is likely to nothing for confirmatory data analysis to consider.}{John Tukey, Exploratory Data Analysis, 1977, p. 3.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [19] "\\epigraph{One thing the data analyst has to learn is how to expose himself to what his data are willing--or even anxious--to tell him. Finding clues requires looking in the right places and with the right magnifying glass.}{John Tukey, Exploratory Data Analysis, 1977, p. 21.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [20] "\\epigraph{In data analysis, a plot of y against x may help us when we know nothing about the logical connection from x to y--even when we do not know whether or not there is one--even when we know that such a connection is impossible.}{John Tukey, Exploratory Data Analysis, p. 131.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [21] "\\epigraph{Whatever the data, we can try to gain understanding by straightening or by flattening. When we succeed in doing one or both, we almost always see more clearly what is going on.}{John Tukey, Exploratory Data Analysis, 1977, p. 148.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [22] "\\epigraph{A competent data analysis of an even moderately complex set of data is a thing of trials and retreats, of dead ends and branches.}{John Tukey, Computer Science and Statistics: Proceedings of the 14th Symposium on the Interface, p. 4.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [23] "\\epigraph{There is no more reason to expect one graph to 'tell all' than to expect one number to do the same.}{John Tukey, Exploratory Data Analysis, 1977, p. 157}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [24] "\\epigraph{There is no excuse for failing to plot and look.}{John Tukey, Exploratory Data Analysis, 1997, p. 43}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [25] "\\epigraph{One is so much less than two. [John Tukey's eulogy of his wife.]}{John Tukey, The life and professional contributions of John W. Tukey, The Annals of Statistics, 2001, Vol 30, p. 46.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [26] "\\epigraph{No one has ever shown that he or she had a free lunch. Here, of course,``free lunch\" means``usefulness of a model that is locally easy to make inferences from\".}{John Tukey, Issues relevant to an honest account of data-based inference, partially in the light of Laurie Davies' paper.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [27] "\\epigraph{If asymptotics are of any real value, it must be because they teach us something useful in finite samples. I wish I knew how to be sure when this happens.}{John Tukey, Issues relevant to an honest account of data-based inference, partially in the light of Laurie Davies' paper.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> [28] "\\epigraph{Statistics is a science in my opinion, and it is no more a branch of mathematics than are physics, chemistry, and economics; for if its methods fail the test of experience--not the test of logic--they will be discarded.}{John Tukey, The life and professional contributions of John W. Tukey, by David Brillinger, The Annals of Statistics, 2001, Vol 30.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [29] "\\epigraph{One Christmas Tukey gave his students books of crossword puzzles as presents. Upon examining the books the students found that Tukey had removed the puzzle answers and had replaced them with words of the sense:``Doing statistics is like doing crosswords except that one cannot know for sure whether one has found the solution.\"}{John Tukey, The life and professional contributions of John W. Tukey, by David Brillinger, The Annals of Statistics, 2001, Vol 30, p. 22.}\n\n"                                                                                                                                                                                                                                                                                                            
#> [30] "\\epigraph{A sort of question that is inevitable is:``Someone taught my students exploratory, and now (boo hoo) they want me to tell them how to assess significance or confidence for all these unusual functions of the data. Oh, what can we do?\" To this there is an easy answer: TEACH them the JACKKNIFE.}{John Tukey, We Need Both Exploratory and Confirmatory, The American Statistician, Vol 34, No 1, p. 25.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                  
#> [31] "\\epigraph{John Tukey's eye for detail was amazing. When we were preparing some of the material for our book (which was published last year), it was most disconcerting to have him glance at the data and question one value out of several thousand points. Of course, he was correct and I had missed identifying this anomaly.}{Kaye Basford, NA}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [32] "\\epigraph{Many students are curious about the '1.5 x IQR Rule';, i.e. why do we use Q1 - 1.5 x IQR (or Q3 + 1.5 x IQR) as the value for deciding if a data value is classified as an outlier? Paul Velleman, a statistician at Cornell University, was a student of John Tukey, who invented the boxplot and the 1.5 x IQR Rule. When he asked Tukey, 'Why 1.5?', Tukey answered, 'Because 1 is too small and 2 is too large.' [Assuming a Gaussian distribution, about 1 value in 100 would be an outlier. Using 2 x IQR would lead to 1 value in 1000 being an outlier.]}{Unknown, NA}\n\n"                                                                                                                                                                                                                  
#> [33] "\\epigraph{It is a rare thing that a specific body of data tells us as clearly as we would wish how it itself should be analyzed.}{John Tukey, Exploratory Data Analysis, 1977, p. 397.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [34] "\\epigraph{Just which robust/resistant methods you use is not important--what is important is that you use some. It is perfectly proper to use both classical and robust/resistant methods routinely, and only worry when they differ enough to matter. But, when they differ, you should think hard.}{John Tukey, Quoted by Doug Martin}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [35] "\\epigraph{I believe that there are many classes of problems where Bayesian analyses are reasonable, mainly classes with which I have little acquaintance.}{John Tukey, The life and professional contributions of John W. Tukey, The Annals of Statistics, 2001, Vol 30, p. 45.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
#> [36] "\\epigraph{The twin assumptions of normality of distribution and homogeneity of variance are not ever exactly fulfilled in practice, and often they do not even hold to a good approximation.}{John W. Tukey, The problem of multiple comparisons. 1973. Unpublished manuscript, Dept. of Statistics, Princeton University.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [37] "\\epigraph{If we need a short suggestion of what exploratory data analysis is, I would suggest that: 1. it is an attitude, AND 2. a flexibility, AND 3. some graph paper (or transparencies, or both).}{John W. Tukey, Jones, L. V. (Ed.). (1986). The collected works of John W. Tukey: Philosophy and principles of data analysis 1949-1964 (Vols. III & IV). London: Chapman & Hall.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [38] "\\epigraph{Three of the main strategies of data analysis are: 1. graphical presentation. 2. provision of flexibility in viewpoint and in facilities,  3. intensive search for parsimony and simplicity.}{John W. Tukey, Jones, L. V. (Ed.). (1986). The collected works of John W. Tukey: Philosophy and principles of data analysis 1949-1964 (Vols. III & IV). London: Chapman & Hall.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                  
#> [39] "\\epigraph{The purpose of plotting is to convey phenomena to the viewer's cortex, not to provide a place to lookup observed numbers.}{Kaye Basford, John Tukey, Graphical Analysis of Multi-Response Data, p. 373.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        
#> [40] "\\epigraph{Had we started with this [quantile] plot, noticed that it looks straight and not looked further, we would have missed the important features of the data.  The general lesson is important. Theoretical quantile-quantile plots are not a panacea and must be used in conjunction with other displays and analyses to get a full picture of the behavior of the data.}{John M. Chambers, William S. Cleveland, Beat Kleiner, Paul A. Tukey, Graphical Methods for Data Analysis, p. 212.}\n\n"                                                                                                                                                                                                                                                                                                       
#> [41] "\\epigraph{Our conclusion about [choropleth] patch maps agrees with Tukey's (1979), who left little doubt about his opinions by stating, 'I am coming to be less and less satisfied with the set of maps that some dignify by the name \\emph{statistical map} and that I would gladly revile with the name \\emph{patch map}'.}{William Cleveland & Robert McGill, Graphical Perception: Theory, Experimentation, and Application to the Development of Graphical Models, Journal of the American Statistical Association, 79, 531--554, 1984.}\n\n"                                                                                                                                                                                                                                                           
#> [42] "\\epigraph{Spatial patterns may be due to many sources of variation. In the context of seeking explanations, John Tukey said that,``the unadjusted plot should not be made.\" In other words, our perceptual/cognitive abilities are poor in terms of adjusting for known source of variations and envisioning the resulting map. A better strategy is to control for known sources of variation and/or adjust the estimates before making the map.}{Dan Carr, Survey Research Methods Section newsletter, July 2002.}\n\n"                                                                                                                                                                                                                                                                                     
#> [43] "\\epigraph{George Box: We don't need robust methods. A good statistician (particularly a Bayesian one) will model the data well and find the outliers.\nJohn Tukey: They ran over 2000 statistical analyses at Rothamsted last week and nobody noticed anything. A red light warning would be most helpful.}{George Box vs. John Tukey, Douglas Martin, 1999 S-Plus Conference Proceedings.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                               
#> [44] "\\epigraph{If significance tests are required for still larger samples, graphical accuracy is insufficient, and arithmetical methods are advised. A word to the wise is in order here, however. Almost never does it make sense to use exact binomial significance tests on such data - for the inevitable small deviations from the mathematical model of independence and constant split have piled up to such an extent that the binomial variability is deeply buried and unnoticeable. Graphical treatment of such large samples may still be worthwhile because it brings the results more vividly to the eye.}{Frederick Mosteller & John W Tukey, The Uses and Usefulness of Binomial Probability Paper, Journal of the American Statistical Association 44, 1949.}\n\n"                                
#> [45] "\\epigraph{Is statistics a science?...If statistics is a science, then what is its subject matter? Physicists study``stuff\", biologists study life, astronomers study the universe—what do statisticians study?\ndoi: 10.1198/004017006000000219}{Colin Mallows, Tukey's Paper After 40 Years, Technometrics, 2006, V 48, No 3. Page 322.}\n\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [46] "\\epigraph{David Cox was asked to identify``what is statistics.\" His answer was that statistics is the discipline concerned with the study of variability, the study of uncertainty, and the study of decision making in the face of uncertainty. ... This seems to say that what statisticians study is their\nmethodology, divorced from applications.\ndoi: 10.1198/004017006000000219}{Colin Mallows, Tukey's Paper After 40 Years, Technometrics, 2006, V 48, No 3. Page 322.}\n\n"                                                                                                                                                                                                                                                                                                                       
#> [47] "\\epigraph{The definition [of statistics] I prefer says that: Statistics concerns the relation of quantitative data to a real-world problem, often in the presence of variability and uncertainty. It attempts to make precise and explicit what the data has to say about the problem of interest.\ndoi: 10.1198/004017006000000219}{Colin Mallows, Tukey's Paper After 40 Years, Technometrics, 2006, V 48, No 3. Page 322.}\n\n"                                                                                                                                                                                                                                                                                                                                                                             
#> [48] "\\epigraph{In 1995 Daryl Pregibon and I gave a definition of a massive-data problem. This is not just a problem for which we have massive data; that could be simply a problem where classical asymptotic theory is all that is needed. I have not seen many of those. No, a massive data problem is a massive problem for which we have massive data.\ndoi: 10.1198/004017006000000219}{Colin Mallows, Tukey's Paper After 40 Years, Technometrics, 2006, V 48, No 3. Page 323.}\n\n"