Display a randomly chosen statistical quote.

Display a randomly chosen statistical quote.

Usage

statquote(ind = NULL, pattern = NULL, tag = NULL, source = NULL, topic = NULL)

# S3 method for class 'statquote'
print(x, cite = TRUE, width = NULL, ...)

# S3 method for class 'statquote'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

Arguments

ind

Integer or character. If 'ind' is missing, a random quote is chosen from all quotations. If 'ind' is specified and is an integer, return the ind^th quote. If 'ind' is specified and is character, use it as the 'pattern'.

pattern

Character string. Quotes are subset to to those which match the pattern in the quote text.

tag

Character string. Quotes are subset to those matching the specified tag.

source

Character string. Quotes are subset to those matching the specified source (person).

topic

Deprecated. Use 'tag' instead. Only kept for backward compatibility.

x

object of class 'statquote'

cite

logical; should the cite field be printed?

width

Optional print width parameter

...

Other optional arguments, unused here

row.names

see as.data.frame

optional

see as.data.frame

Value

A character vector containing one quote. It is of class statquote for which an S3 print method will be invoked, and for which other methods are available.

See also

Examples

set.seed(1234)
statquote()
#> 
#> Stepwise regression is probably the most abused computerized
#> statistical technique ever devised.  If you think you need stepwise
#> regression to solve a particular problem you have, it is almost certain
#> that you do not. Professional statisticians rarely use automated
#> stepwise regression.
#> --- Leland Wilkinson, SYSTAT (1984). P. 196.
statquote(10)
#> 
#> The physical sciences are used to 'praying over' their data, examining
#> the same data from a variety of points of view. This process has been
#> very rewarding, and has led to many extremely valuable insights.
#> Without this sort of flexibility, progress in physical science would
#> have been much slower. Flexibility in analysis is often to be had
#> honestly at the price of a willingness not to demand that what has
#> already been observed shall establish, or prove, what analysis
#> suggests. In physical science generally, the results of praying over
#> the data are thought of as something to be put to further test in
#> another experiment, as indications rather than conclusions.
#> --- John Tukey, The Future of Data Analysis, Annals of Mathematical
#> Statistics, 1962, Vol. 33 (1), 1962.
statquote("boggled")
#> 
#> The statistician has no magic touch by which he may come in at the
#> stage of tabulation and make something of nothing. Neither will his
#> advice, however wise in the early stages of a study, ensure successful
#> execution and conclusion. Many a study, launched on the ways of elegant
#> statistical design, later boggled in execution, ends up with results to
#> which the theory of probability can contribute little.
#> --- W. Edwards Deming, Principles of Professional Statistical Practice.
#> Annals of Mathematical Statistics, 36(6), 1883. (1965)
statquote(pattern="boggled")
#> 
#> The statistician has no magic touch by which he may come in at the
#> stage of tabulation and make something of nothing. Neither will his
#> advice, however wise in the early stages of a study, ensure successful
#> execution and conclusion. Many a study, launched on the ways of elegant
#> statistical design, later boggled in execution, ends up with results to
#> which the theory of probability can contribute little.
#> --- W. Edwards Deming, Principles of Professional Statistical Practice.
#> Annals of Mathematical Statistics, 36(6), 1883. (1965)
statquote(source="Yates")
#> 
#> [Researchers] pay undue attention to the results of tests of
#> significance they perform on their data, particularly data derived from
#> experiments, and too little to the estimates of the magnitude of the
#> effects which they are investigating.... The emphasis on tests of
#> significance, and the consideration of the results of each experiment
#> in isolation, have had the unfortunate consequence that scientific
#> workers have often regarded the execution of a test of significance on
#> an experiment as the ultimate objective. Results are significant or not
#> and that is the end to it.
#> --- Frank Yates, The influence of Statistical Methods for Research
#> Workers on the development of the science of statistics. 1951. Journal
#> of the American Statistical Association 46: 19-34.
statquote(tag="anova")
#> 
#> the number of stars by itself is relevant only to the question of
#> whether H0 is exactly true--a question which is almost always not of
#> interest to us, especially because we usually know a priori that H0
#> cannot be exactly true.
#> --- Norman S. Matloff, Statistical hypothesis testing: problems and
#> alternatives. 1991. Environmental Entomology 20 : 1246-1250.
print.data.frame(statquote(302)) # All information
#>     qid                                                                 text
#> 302 302 It is a part of probability that many improbable things will happen.
#>                    source             cite  url        tags  tex
#> 302 Agathon, 445 - 400 BC Chance News 7.02 <NA> probability <NA>