Function coerces statquote objects to strings suitable for LaTeX

This function coerces statquote objects to strings suitable for rendering in LaTeX. Quotes and (potential LaTeX) sources are placed within suitable "epigraph" output format via the sprintf function.

as.latex(quotes, form = "\\epigraph{%s}{%s}\n\n")

Arguments

quotes
an object of class statquote returned from functions such as search_quotes or statquote

form
structure of the LaTeX output for the text (first argument) and source (second argument) passed to sprintf

Value

character vector of formatted LaTeX quotes

Examples

ll <- search_quotes("Tukey")
as.latex(ll)

#>  [1] "\\epigraph{The greatest value of a picture is when it forces us to notice what we never expected to see.}{John W. Tukey \\citet[p. vi]{Tukey:77}}\n\n"                                                                                                                                                                                                                                                                                          
#>  [2] "\\epigraph{If one technique of data analysis were to be exalted above all others for its ability to be revealing to the mind in connection with each of many different models, there is little doubt which one would be chosen. The simple graph has brought more information to the data analyst's mind than any other device. It specializes in providing indications of unexpected phenomena.}{John W. Tukey \\citet[p. 49]{Tukey:1962}}\n\n"
#>  [3] "\\epigraph{The greatest possibilities of visual display lie in vividness and inescapability of the intended message. A visual display can stop your mental flow in its tracks and make you think. A visual display can force you to notice what you never expected to see.}{John W. Tukey \\citeyear{Tukey:90}}\n\n"                                                                                                                            
#>  [4] "\\epigraph{The purpose of [data] display is comparison (recognition of phenomena), not numbers ... The phenomena are the main actors, numbers are the supporting cast.}{John W. Tukey \\citeyear{Tukey:90}}\n\n"                                                                                                                                                                                                                                
#>  [5] "\\epigraph{...But it is not always clear \\emph{which} 1000 words.}{John W. Tukey, 1973}\n\n"                                                                                                                                                                                                                                                                                                                                                   
#>  [6] "\\epigraph{Exploratory data analysis can never be the whole story, but nothing else\tcan serve as the foundation stone -- as the first step.}{John W. Tukey, EDA, 1977, p.3.}\n\n"                                                                                                                                                                                                                                                               
#>  [7] "\\epigraph{The best thing about being a statistician is that you get to play in everyone's backyard.}{John W. Tukey}\n\n"                                                                                                                                                                                                                                                                                                                       
#>  [8] "\\epigraph{Far better an approximate answer to the right question, which is often vague, than an exact answer to the wrong question, which can always be made precise.}{John W. Tukey, (1962),``The future of data analysis,\" Annals of Mathematical Statistics 33, 1-67.}\n\n"                                                                                                                                                                
#>  [9] "\\epigraph{A bad answer to a good question may be far better than a good answer to a bad question.}{A graduate class, paraphrasing Tukeys dictum.}\n\n"                                                                                                                                                                                                                                                                                         
#> [10] "\\epigraph{The worst, i.e., most dangerous, feature of 'accepting the null hypothesis' is the giving up of explicit uncertainty ... Mathematics can sometimes be put in such black-and-white terms, but our knowledge or belief about the external world never can.}{John W. Tukey. (1991).``The Philosophy of Multiple Comparisons,\" Statistical Science 6, 100--116.}\n\n"                                                                   
#> [11] "\\epigraph{Better to have an approximate answer to the right question than a precise answer to the wrong question.}{John W. Tukey as quoted by John Chambers}\n\n"                                                                                                                                                                                                                                                                              
#> [12] "\\epigraph{The combination of some data and an aching desire for an answer does not ensure that a reasonable answer can be extracted from a given body of data.}{John W. Tukey}\n\n"                                                                                                                                                                                                                                                            
#> [13] "\\epigraph{The practical power of a statistical test is the product of its' statistical power and the probability of use.}{John W. Tukey, 1959 \\citep{Tukey:59}}\n\n"                                                                                                                                                                                                                                                                          
#> [14] "\\epigraph{The best thing about being a statistician is that you get to play in everyone's backyard.}{John W. Tukey}\n\n"